domain int 형 -> long 형으로 변경

+++ src/main/java/com/nhnacademy/midoo/gateway/domain/AccountIdOnly.java
@@ -4,5 +4,5 @@ import lombok.Value;
 
 @Value
 public class AccountIdOnly {
-    int accountId;
+    long accountId;
 }
