+++ src/main/java/com/nhnacademy/midoo/gateway/domain/TagPostRequest.java
@@ -5,5 +5,5 @@ import lombok.Value;
 @Value
 public class TagPostRequest {
     String tagName;
-    int projectId;
+    long projectId;
 }
